+++ src/Expressions/Matches/Conditional.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2023-2026 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/rector-rules
+ */
+
+namespace Ergebnis\Rector\Rules\Expressions\Matches;
+
+/**
+ * @internal
+ */
+interface Conditional
+{
+}