+++ src/Expressions/Matches/IntConditional.php
@@ -1,0 +1,37 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2023-2026 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/rector-rules
+ */
+
+namespace Ergebnis\Rector\Rules\Expressions\Matches;
+
+/**
+ * @internal
+ */
+final class IntConditional implements Conditional
+{
+    private int $value;
+
+    private function __construct(int $value)
+    {
+        $this->value = $value;
+    }
+
+    public static function fromInt(int $value): self
+    {
+        return new self($value);
+    }
+
+    public function toInt(): int
+    {
+        return $this->value;
+    }
+}